LayoutTests/js/logical-or-jless-expected.txt

This test checks whether the pair of opcodes (less, jtrue) is incorrectly optimized in a LogicalOrNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is true
PASS successfullyParsed is true

TEST COMPLETE
